fix: remove Clang compilation errors

+++ src/snsfopt.cpp
@@ -127,7 +127,7 @@ std::string SnsfOpt::ToTimeString(double t, bool padding)
 	if (!padding)
 	{
 		size_t len = strlen(str);
-		for (size_t i = len - 1; i >= 0; i--)
+		for (off_t i = (off_t)(len - 1); i >= 0; i--)
 		{
 			if (str[i] == '0')
 			{
@@ -1656,10 +1656,10 @@ int main(int argc, char *argv[])
 				printf("Optimizing %s  Song value %X\n", argv[argi], song);
 
 				uint8_t patch[4] = {
-					song & 0xff,
-					(song >> 8) & 0xff,
-					(song >> 16) & 0xff,
-					(song >> 24) & 0xff,
+					static_cast<uint8_t>(song & 0xff),
+					static_cast<uint8_t>((song >> 8) & 0xff),
+					static_cast<uint8_t>((song >> 16) & 0xff),
+					static_cast<uint8_t>((song >> 24) & 0xff),
 				};
 				opt.PatchROM(minisnsf_offset, patch, minisnsf_size, true);
 				opt.ResetGame();
